fix(sdk-coin-sui): return all input objects in `toJson` result
Ticket: WIN-3242
