+++ modules/sdk-coin-sui/src/lib/tokenTransferTransaction.ts
@@ -11,9 +11,15 @@ import {
   TransactionRecipient,
   TransactionType,
 } from '@bitgo/sdk-core';
-import { CallArg, normalizeSuiAddress } from './mystenlab/types';
-import utils from './utils';
-import { builder, Inputs, TransactionBlockInput } from './mystenlab/builder';
+import { CallArg, normalizeSuiAddress, SuiObjectRef } from './mystenlab/types';
+import utils, { isImmOrOwnedObj } from './utils';
+import {
+  builder,
+  Inputs,
+  TransactionArgument,
+  TransactionBlockInput,
+  TransactionType as SuiTransactionBlockType,
+} from './mystenlab/builder';
 import { BCS } from '@mysten/bcs';
 
 export class TokenTransferTransaction extends Transaction<TokenTransferProgrammableTransaction> {
@@ -70,6 +76,7 @@ export class TokenTransferTransaction extends Transaction<TokenTransferProgramma
       kind: { ProgrammableTransaction: tx.tx },
       gasData: tx.gasData,
       expiration: { None: null },
+      inputObjects: this.getInputObjectsFromTx(tx.tx),
     };
   }
 
@@ -156,7 +163,6 @@ export class TokenTransferTransaction extends Transaction<TokenTransferProgramma
    * @return {TxData}
    */
   public getTxData(): TxData {
-    // TODO
     if (!this._suiTransaction) {
       throw new InvalidTransactionError('empty transaction');
     }
@@ -212,5 +218,38 @@ export class TokenTransferTransaction extends Transaction<TokenTransferProgramma
       outputAmount,
       outputs,
     };
+  }
+
+  /**
+   * Extracts the objects that were provided as inputs while building the transaction
+   * @param tx
+   * @returns {SuiObjectRef[]} Objects that are inputs for the transaction
+   */
+  private getInputObjectsFromTx(tx: TokenTransferProgrammableTransaction): SuiObjectRef[] {
+    const inputs = tx.inputs;
+    const transaction = tx.transactions[0] as SuiTransactionBlockType;
+
+    let args: TransactionArgument[] = [];
+    if (transaction.kind === 'MergeCoins') {
+      const { destination, sources } = transaction;
+      args = [destination, ...sources];
+    } else if (transaction.kind === 'SplitCoins') {
+      args = [transaction.coin];
+    }
+
+    const inputObjects: SuiObjectRef[] = [];
+    args.forEach((arg) => {
+      if (arg.kind === 'Input') {
+        let input = inputs[arg.index];
+        if ('value' in input) {
+          input = input.value;
+        }
+        if ('Object' in input && isImmOrOwnedObj(input.Object)) {
+          inputObjects.push(input.Object.ImmOrOwned);
+        }
+      }
+    });
+
+    return inputObjects;
   }
 }
