+++ modules/sdk-coin-sui/test/unit/transactionBuilder/tokenTransferBuilder.ts
@@ -74,6 +74,8 @@ describe('Sui Token Transfer Builder', () => {
       rebuilder.addSignature({ pub: testData.sender.publicKey }, Buffer.from(testData.sender.signatureHex));
       const rebuiltTx = await rebuilder.build();
       rebuiltTx.toBroadcastFormat().should.equal(rawTx);
+      rebuiltTx.toJson().gasData.payment.length.should.equal(numberOfGasPaymentObjects);
+      rebuiltTx.toJson().inputObjects.length.should.equal(numberOfInputObjects);
     });
   });
