+++ modules/sdk-coin-sui/test/unit/transactionBuilder/transferBuilder.ts
@@ -141,6 +141,7 @@ describe('Sui Transfer Builder', () => {
       rebuilder.addSignature({ pub: testData.sender.publicKey }, Buffer.from(testData.sender.signatureHex));
       const rebuiltTx = await rebuilder.build();
       rebuiltTx.toBroadcastFormat().should.equal(rawTx);
+      rebuiltTx.toJson().gasData.payment.length.should.equal(numberOfPaymentObjects);
     });
   });
